backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
Pillow==10.1.0
python-decouple==3.8
boto3==1.34.0
django-storages==1.14.2
localstack==3.0.0
